+++ back/scada-back/Migrations/ScadaDbContextModelSnapshot.cs
@@ -134,8 +134,8 @@ namespace scada_back.Migrations
                         .HasColumnType("float")
                         .HasColumnName("high_limit");
 
-                    b.Property<int>("InitialValue")
-                        .HasColumnType("int")
+                    b.Property<double>("InitialValue")
+                        .HasColumnType("float")
                         .HasColumnName("initial_value");
 
                     b.Property<double>("LowLimit")
@@ -221,8 +221,8 @@ namespace scada_back.Migrations
                         .HasColumnType("nvarchar(max)")
                         .HasColumnName("description");
 
-                    b.Property<int>("InitialValue")
-                        .HasColumnType("int")
+                    b.Property<double>("InitialValue")
+                        .HasColumnType("float")
                         .HasColumnName("initial_value");
 
                     b.Property<string>("Name")
