+++ back/scada-back/Models/DigitalOutput.cs
@@ -19,6 +19,6 @@ namespace scada_back.Models
         public string Address { get; set; }
 
         [Column("initial_value")]
-        public int InitialValue { get; set; }
+        public double InitialValue { get; set; }
     }
 }
